+++ application/modules/gejala/views/datagrid.php
@@ -16,8 +16,10 @@
                         <thead>
                             <tr>
                                 <th class="col-lg-2">ID</th>
-                                <th>Gejala</th>
-                                <th class="col-lg-1">Aksi</th>
+                                <th>Nama</th>
+                                <?php if($this->session->userdata('id')) { ?>
+                            		<th class="col-lg-1">Aksi</th>    	
+                                <?php } ?>
                             </tr>
                         </thead>
                         <tbody>
@@ -25,13 +27,15 @@
                             <tr class="odd gradeX">
 								<td><?php echo $row->id; ?></td>
                                 <td><?php echo $row->gejala; ?></td>
-                                <td class="center">
-									<?php
-                            			echo anchor('gejala/form_update/'.$row->id, '<i class="fa fa-edit"></i>', 'class="btn btn-xs" title="Edit"');
-                            			echo '<a href="'.base_url().'gejala/delete_gejala/'.$row->id.'" onclick="return confirmModal(\'Anda akan menghapus?\',\''.base_url().'gejala/delete_gejala/'.$row->id.'\')" class="btn btn-xs" title="Hapus"><i class="fa fa-trash-o"></i></a>'; 
-									?>
-								</td>
-							</tr>
+                                <?php if($this->session->userdata('id')) { ?>
+                            		<td class="center">
+	                                	<?php
+											echo anchor('gejala/form_update/'.$row->id, '<i class="fa fa-edit"></i>', 'class="btn btn-xs" title="Edit"');
+    	                        			echo '<a href="'.base_url().'gejala/delete_gejala/'.$row->id.'" onclick="return confirmModal(\'Anda akan menghapus?\',\''.base_url().'gejala/delete_gejala/'.$row->id.'\')" class="btn btn-xs" title="Hapus"><i class="fa fa-trash-o"></i></a>';	
+										?>
+									</td>    	
+                                <?php } ?>
+                            </tr>
 							<?php endforeach; ?>
                  		</tbody>
                     </table>
