+++ application/modules/penyakit/views/datagrid.php
@@ -26,10 +26,17 @@
 								<td><?php echo $row->id; ?></td>
                                 <td><?php echo $row->penyakit; ?></td>
                                 <td class="center">
-									<?php
-                            			echo anchor('penyakit/form_update/'.$row->id, '<i class="fa fa-edit"></i>', 'class="btn btn-xs" title="Edit"');
-                            			echo anchor('penyakit/detail_penyakit/'.$row->id, '<i class="fa fa-eye"></i>', 'class="btn btn-xs" title="Detail"');
-                            			echo '<a href="'.base_url().'penyakit/delete_penyakit/'.$row->id.'" onclick="return confirmModal(\'Anda akan menghapus?\',\''.base_url().'penyakit/delete_penyakit/'.$row->id.'\')" class="btn btn-xs" title="Hapus"><i class="fa fa-trash-o"></i></a>'; 
+                                	<?php
+										if (!$this->session->userdata('id'))
+										{
+											echo anchor('penyakit/detail_penyakit/'.$row->id, '<i class="fa fa-eye"></i>', 'class="btn btn-xs" title="Detail"');
+										}
+										else
+										{
+											echo anchor('penyakit/form_update/'.$row->id, '<i class="fa fa-edit"></i>', 'class="btn btn-xs" title="Edit"');
+                            				echo anchor('penyakit/detail_penyakit/'.$row->id, '<i class="fa fa-eye"></i>', 'class="btn btn-xs" title="Detail"');
+                            				echo '<a href="'.base_url().'penyakit/delete_penyakit/'.$row->id.'" onclick="return confirmModal(\'Anda akan menghapus?\',\''.base_url().'penyakit/delete_penyakit/'.$row->id.'\')" class="btn btn-xs" title="Hapus"><i class="fa fa-trash-o"></i></a>';	
+										}
 									?>
 								</td>
 							</tr>
