+++ application/modules/solusi/views/datagrid.php
@@ -17,7 +17,9 @@
                             <tr>
                                 <th class="col-lg-1">No</th>
                                 <th>Solusi</th>
-								<th class="col-lg-1">Aksi</th>
+                                <?php if($this->session->userdata('id')) { ?>
+                            		<th class="col-lg-1">Aksi</th>    	
+                                <?php } ?>
                             </tr>
                         </thead>
                         <tbody>
@@ -25,12 +27,14 @@
                             <tr class="odd gradeX">
                                 <td>&nbsp;</td>
                                 <td><?php echo $row->solusi; ?></td>
-                                <td class="center col-lg-1">
-                                   	<?php
-                            			echo anchor('solusi/form/'.$row->id, '<i class="fa fa-edit"></i>', 'class="btn btn-xs" title="Edit"');
-                            			echo '<a href="'.base_url().'solusi/delete_tritmen/'.$row->id.'" onclick="return confirmModal(\'Anda akan menghapus?\',\''.base_url().'solusi/delete_tritmen/'.$row->id.'\')" class="btn btn-xs" title="Hapus"><i class="fa fa-trash-o"></i></a>'; 
-									?>
-                                </td>
+                                <?php if($this->session->userdata('id')) { ?>
+                            		<td class="center">
+	                                	<?php
+											echo anchor('solusi/form/'.$row->id, '<i class="fa fa-edit"></i>', 'class="btn btn-xs" title="Edit"');
+                            			echo '<a href="'.base_url().'solusi/delete_tritmen/'.$row->id.'" onclick="return confirmModal(\'Anda akan menghapus?\',\''.base_url().'solusi/delete_tritmen/'.$row->id.'\')" class="btn btn-xs" title="Hapus"><i class="fa fa-trash-o"></i></a>';	
+										?>
+									</td>    	
+                                <?php } ?>
 							</tr>
 							<?php endforeach; ?>
                  		</tbody>
